requirements.txt
mongoengine
fastapi
uvicorn
requests
pyyaml
AWSIoTPythonSDK
python-multipart
cantools
influxdb
boto3